plotting things and crying

+++ examples/eval_scripts/view_intergen_model.py
@@ -61,7 +61,7 @@ def main(roles,
   level_name = get_name_from_rules(rules)
   substrate_name = f'rule_obeying_harvest_{level_name}'
   num_bots = len(roles)
-  age_range = DEFAULT_MAX_LIFE_SPAN / len(roles)
+  age_range = int(DEFAULT_MAX_LIFE_SPAN / len(roles))
   ages = list(range(10, len(roles)*age_range, age_range)) + [0]
   random.shuffle(ages) # that not nearby agents a likely to die
